+++ src/main/java/com/lvs/profiler/service/ClickService.java
@@ -9,9 +9,9 @@ import java.util.Date;
 
 public interface ClickService {
 
-    SearchResponse findByUserHotel(int uId, int topCount);
+    SearchResponse findByUserHotel(Long uId, int topCount);
 
-    SearchResponse findByUserHotelRegion(int uId, int topCount);
+    SearchResponse findByUserHotelRegion(Long uId, int topCount);
 
     Page<Click> findAllClick(Pageable pageRequest);
